requirements.txt
certifi==2025.6.15+computecanada
charset-normalizer==3.4.2
contourpy==1.3.1+computecanada
cycler==0.12.1+computecanada
fairlearn==0.12.0
filelock==3.18.0+computecanada
fonttools==4.59.0+computecanada
fsspec==2025.5.1+computecanada
hf_xet==1.1.3+computecanada
huggingface-hub==0.33.0
idna==3.10+computecanada
jinja2==3.1.6+computecanada
joblib==1.5.1+computecanada
kiwisolver==1.4.8+computecanada
MarkupSafe==2.1.5+computecanada
matplotlib==3.10.1+computecanada
mpmath==1.3.0+computecanada
networkx==3.5+computecanada
numpy==2.2.2+computecanada
packaging==25.0+computecanada
pandas==2.2.3+computecanada
pillow==11.1.0+computecanada
pyparsing==3.2.3+computecanada
python_dateutil==2.9.0.post0+computecanada
pytz==2025.2+computecanada
PyYAML==6.0.2+computecanada
regex==2024.11.6+computecanada
requests==2.32.4+computecanada
safetensors==0.5.3+computecanada
scikit_learn==1.6.1+computecanada
scipy==1.15.1+computecanada
seaborn==0.13.2+computecanada
sentence-transformers==4.1.0
six==1.17.0+computecanada
sklearn==0.0+computecanada
sympy==1.14.0+computecanada
threadpoolctl==3.6.0+computecanada
tokenizers==0.21.1+computecanada
torch==2.7.1+computecanada
tqdm==4.67.1+computecanada
transformers==4.52.4+computecanada
typing_extensions==4.14.0+computecanada
tzdata==2025.2+computecanada
urllib3==2.5.0+computecanada
